Starting: adding path_requests_run functions to the "core" package
- Path_request class is updated to be more generic and reuseable
  in transmission main example. json processing linked to the yang
  modelling is kept in the pat_request_run, while path_request class
  only contains useful attributes
- adding functions in info and request python files in core directory
