+++ examples/path_requests_run.py
@@ -30,7 +30,8 @@ from gnpy.core.network import load_network, build_network
 from gnpy.core.equipment import load_equipment
 from gnpy.core.elements import Transceiver, Roadm, Edfa, Fused
 from gnpy.core.utils import db2lin, lin2db
-from gnpy.core.info import create_input_spectral_information, SpectralInformation, Channel, Power
+from gnpy.core.info import create_input_spectral_information, SpectralInformation, Channel, Power, load_SI
+from gnpy.core.request import Path_request, Result_element
 from copy import copy, deepcopy
 from numpy import log10
 
@@ -46,114 +47,6 @@ parser.add_argument('-v', '--verbose', action='count')
 parser.add_argument('-o', '--output', default=None)
 
 
-class Path_request():
-    def __init__(self,jsondata,tspjsondata):
-        self.request_id = jsondata['request-id']
-        self.source = jsondata['src-tp-id']
-        self.destination = jsondata['dst-tp-id']
-        # retrieving baudrate out of transponder type and mode (format)
-        self.tsp = jsondata['path-constraints']['te-bandwidth']['trx_type']
-        self.tsp_mode = jsondata['path-constraints']['te-bandwidth']['trx_mode']
-        # for debug
-        # print(tsp)
-        try:
-            baudrate = next(m['baudrate'] 
-                for t in  tspjsondata if t['type_variety'] == self.tsp
-                for m in t['mode']  if  m['format'] == self.tsp_mode)
-        except StopIteration:
-            msg = f'could not find tsp : {self.tsp} with mode: {self.tsp_mode} in eqpt library'
-            logger.critical(msg)
-            raise ValueError(msg)
-        self.baudrate = baudrate
-
-        nodes_list = jsondata['optimizations']['explicit-route-include-objects']
-        self.nodes_list = [n['unnumbered-hop']['node-id'] for n in nodes_list]
-        # create a list for individual loose capability for each node ... 
-        # even if convert_service_sheet fills it with the same value
-        self.loose_list = [n['unnumbered-hop']['hop-type'] for n in nodes_list]
-
-        self.spacing = jsondata['path-constraints']['te-bandwidth']['spacing']
-        self.power = jsondata['path-constraints']['te-bandwidth']['output-power']
-        self.nb_channel = jsondata['path-constraints']['te-bandwidth']['max-nb-of-channel']
-
-    def __str__(self):
-        return '\n\t'.join([  f'{type(self).__name__} {self.request_id}',
-                            f'source:       {self.source}',
-                            f'destination:  {self.destination}'])
-    def __repr__(self):
-        return '\n\t'.join([  f'{type(self).__name__} {self.request_id}',
-                            f'source:       {self.source}',
-                            f'destination:  {self.destination}',
-                            f'trx type:     {self.tsp}',
-                            f'baudrate:     {self.baudrate}',
-                            f'spacing:      {self.spacing}',
-                            f'power:        {self.power}'
-                            '\n'])
-
-
-class Result_element(Element):
-    def __init__(self,path_request,computed_path):
-        self.path_id = int(path_request.request_id)
-        self.path_request = path_request
-        self.computed_path = computed_path
-        hop_type = []
-        for e in computed_path :
-            if isinstance(e, Transceiver) : 
-                hop_type.append(' - '.join([path_request.tsp,path_request.tsp_mode])) 
-            else:
-                hop_type.append('not recorded')
-        self.hop_type = hop_type
-    uid = property(lambda self: repr(self))
-    @property
-    def pathresult(self):
-        return {
-               'path-id': self.path_id,
-               'path-properties':{
-                   'path-metric': [
-                       {
-                       'metric-type': 'SNR@bandwidth',
-                       'accumulative-value': round(mean(self.computed_path[-1].snr),2)
-                       },
-                       {
-                       'metric-type': 'SNR@0.1nm',
-                       'accumulative-value': round(mean(self.computed_path[-1].snr+10*log10(self.path_request.baudrate/12.5)),2)
-                       }
-                    ],
-                    'path-srlgs': {
-                        'usage': 'not used yet',
-                        'values': 'not used yet'
-                    },
-                    'path-route-objects': [
-                        {
-                        'path-route-object': {
-                            'index': self.computed_path.index(n),
-                            'unnumbered-hop': {
-                                'node-id': n.uid,
-                                'link-tp-id': n.uid,
-                                'hop-type': self.hop_type[self.computed_path.index(n)],
-                                'direction': 'not used'
-                            },
-                            'label-hop': {
-                                'te-label': {
-                                    'generic': 'not used yet',
-                                    'direction': 'not used yet'
-                                    }
-                                }
-                            }
-                        } for n in self.computed_path
-                        ]
-                }
-            }
-                    
-    @property
-    def json(self):
-        return self.pathresult 
-
-def load_SI(filename):
-    with open(filename) as f:
-        json_data = loads(f.read())
-        return json_data['SI'][0]
-
 def load_Transceiver(filename):
     with open(filename) as f:
         json_data = loads(f.read())
@@ -162,9 +55,31 @@ def load_Transceiver(filename):
 def requests_from_json(json_data,eqpt_filename):
     requests_list = []
     tspjsondata = load_Transceiver(eqpt_filename)
+
     for req in json_data['path-request']:
         #print(f'{req}')
-        requests_list.append(Path_request(req,tspjsondata))
+        params = {}
+        params['request_id'] = req['request-id']
+        params['source'] = req['src-tp-id']
+        params['destination'] = req['dst-tp-id']
+        params['trx_type'] = req['path-constraints']['te-bandwidth']['trx_type']
+        params['trx_mode'] = req['path-constraints']['te-bandwidth']['trx_mode']
+        try:
+            extra_params = next(m 
+            for t in  tspjsondata if t['type_variety'] == params['trx_type']
+            for m in t['mode']  if  m['format'] == params['trx_mode'])
+        except StopIteration :
+            msg = f'could not find tsp : {params} with mode: {params} in eqpt library'
+            raise ValueError(msg)
+        nd_list = req['optimizations']['explicit-route-include-objects']
+        params['nodes_list'] = [n['unnumbered-hop']['node-id'] for n in nd_list]
+        params['loose_list'] = [n['unnumbered-hop']['hop-type'] for n in nd_list]
+        params['spacing'] = req['path-constraints']['te-bandwidth']['spacing']
+        params['power'] = req['path-constraints']['te-bandwidth']['output-power']
+        params['nb_channel'] = req['path-constraints']['te-bandwidth']['max-nb-of-channel']
+
+        params.update(extra_params)
+        requests_list.append(Path_request(**params))
 
     return requests_list
